Introduce unique_ptr
Pointers to volatile nodes just allocated are stored
in unique_ptr templates - deallocating the nodes whenever
they go out of scope.

This allows removing the Shutdown method, as manually calling delete
is not needed anymore.

Fixes a potential leak in KVTree::Put -- the leafnode was not
free'ed on transaction abort (missing delete in catch blocks):

```
try {
    ...
    leafnode = new KVLeafNode();
    ...
} catch (nvml::transaction_alloc_error) {
    return FAILED;
} catch (nvml::transaction_error) {
    return FAILED;
}
```

Since that leaf node is stored in a unique_ptr<KVLeafNode> instance, it
is deleted when an exception is raised.

One try-catch block can be removed from KVTree::LeafSplitFull -- as the
leaf node allocated inside it is also automatically deleted once it goes
out of scope, a task which was done by a catch block before:

```
} catch (...) {
    delete new_leafnode;
    throw;
}
```

Fixed an off-by-one bug in KVTree::InnerUpdateAfterSplit

The loop copied the children[idx] element to children[idx + 1] :

`for (int i = keycount; i >= idx; i--) inner->children[i + 1] = inner->children[i];`

With the smart pointer approach this resulted in losing the element
at children[idx] -- since unique_ptr elements can not be copied, they
are moved, clearing children[idx]. When children[idx + 1] is later overwritten,
the destructor of the node at children[idx + 1] is called:

1) `inner->children[idx + 1] = move(inner->children[idx]);` // sets children[idx] to nullptr
2) `inner->children[idx + 1] = move(new_node);` // destroys node at children[idx + 1], before overwriting it

This is fixed by changing the loop condition to `i >= idx`.

The new KVInnerNode::assert_invariants method helped during development.