Adding new engines based on memkind
Adding two new engines for volatile usage of persistent memory.
This two engines built on top of new C++ pmem allocator. This allocator
currently is on review for memkind library. When The C++ pmem allocator
will be accepted to be part of memkind we can enable these two new
engines.

Right now we are commiting only sources, without changes in the build
system.