Use native cmake facility for google test
It's needed to enable using parametrized tests from ctest.
Small drawback is that, user cannot list tests before build.
As homegrown tests collector and gtest downloader is no more needed it
may be removed. Gtest-devel or libgtest-dev is build environment
dependency (look at dockerfiles) so it is not needed to download gtest in runtime
anymore