Remove c api tests for chaching engine
For now those tests should be removed, and added in future if caching
engine is going to be productized. Currently caching and it's tests are
not supported in CI build environment.
To properly implement caching tests, it is needed to add support of reading
json files to test framework, and optionaly to build extended
envionment with own redis instance.