Make the documention build reproducibly
Whilst working on the Reproducible Builds effort [0] we noticed that
pmemkv could not be built reproducibly.

This is because, whilst it used the date in manpages from
SOURCE_DATE_EPOCH [1] (ie. the `dt` variable), it used the current
build year when templating the `year` variable.

This was originally filed in Debian as #952990 [2].

 [0] https://reproducible-builds.org/
 [1] https://reproducible-builds.org/specs/source-date-epoch/
 [2] https://bugs.debian.org/952990