add create_if_missing config's parameter
it allows to create pool file if it doesn't yet exist and it opens it, if it exists.
This flag is mutually exclusive with previously defined config param
create_or_error_if_exists - it may be confusing to set them both since
one would be quietly ignored.

Extend tests for both flags.