tests: use sizeof on type, not on pointer
and add additional, extra padding in custom struct to force fail on such issue.

Found by Klocwork with message:
sizeof should be use on type and not on a pointer. Retrive sizeof
const is usually a developer error that uses the size of the
pointer instead of sizing the type.